carloan/requirements.txt
fastapi
uvicorn
streamlit
pdfplumber
requests
python-dotenv
python-multipart
groq
